flask/requirements.txt
bootstrap-flask
starlette
pydantic
openshift
pyjwt
psycopg2-binary
gunicorn
daiquiri
passgen
passlib
bcrypt
python-multipart
python-dotenv
requests
authlib
six
flask
auth0-python
wtforms
pybreaker